+++ R/keyToEnglish.R
@@ -118,15 +118,27 @@ corpora_to_word_list <- function(
 #' @param hash_subsection_size `numeric` length of each subsection of hash to use for word index. 16^N
 #'                              unique words can be used for a size of N. This value times
 #'                              phrase_length must be less than or equal to the length of the
-#'                              hash output.
+#'                              hash output. Must be less than 14.
 #' @param sep `character` separator to use between each word.
 #' @param word_trans A `function`, `list` of functions, or 'camel' (for CamelCase). If
 #'                   a list is used, then the index of the word of each phrase is
 #'                   mapped to the corresponding function with that index,
 #'                   recycling as necessary
 #'
-#' @param suppress_warning `logical` value indicating if warning of non-character
+#' @param suppress_warnings `logical` value indicating if warning of non-character
 #'                                   input should be suppressed
+#'
+#' @param hash_output_length optional `numeric` if the provided hash function is not a `character`. This is used
+#'                           to send warnings if the hash output is too small to provide full range of all
+#'                           possible combinations of outputs.
+#'
+#' @param forced_limit for multiple word lists, this is the maximum number of values used for calculating the index
+#'                     (prior to taking the modulus) for each word in a phrase. Using this may speed up processing
+#'                     longer word lists with a large least-common-multiple among individual word list lengths. This
+#'                     will introduce a small amount of bias into the randomness. This value should be much larger than
+#'                     any individual word list whose length is not a factor of this value.
+#'
+#' @param numeric_append_range optional `numeric` value of two integers indicating range of integers to append onto data
 #'
 #' @return `character` vector of hashed field resembling phrases
 #'
@@ -145,30 +157,85 @@ keyToEnglish <- function(
   hash_subsection_size=3,
   sep='',
   word_trans='camel',
-  suppress_warning=FALSE
+  suppress_warnings=FALSE,
+  hash_output_length=NA,
+  forced_limit=NA,
+  numeric_append_range=NA
 ){
+  if (hash_subsection_size > 13){
+    stop('hash_subsection_size must not be greater than 13.')
+  }
   KTI_METHOD_LIST=list(
     md5=openssl::md5,
     md4=openssl::md4,
     sha256=openssl::sha256,
+    sha512=openssl::sha512,
+    sha384=openssl::sha384,
     sha1=openssl::sha1,
     sha2=openssl::sha2,
     sha224=openssl::sha224
   )
 
+  KTI_SIZES = list(
+    md5=32,
+    md4=32,
+    sha256=64,
+    sha512=128,
+    sha384=96,
+    sha1=40,
+    sha2=64,
+    sha224=56
+  )
+
+  if (is.character(hash_function)){
+    hash_output_length = KTI_SIZES[[hash_function]]
+  }
+
+  if (is.na(word_list[1])){
+    word_list = utils::read.csv(corpus_path, header=FALSE)[,1,T]
+  }
+
+  if (!is.na(numeric_append_range[1])){
+    if (is.list(word_list)){
+      # nothing special
+    } else {
+      word_list = replicate(phrase_length, word_list, simplify=FALSE)
+    }
+    append = as.character(seq.int(from=numeric_append_range[1], to=numeric_append_range[2]))
+    word_list[[length(word_list)+1]] = append
+  }
+
   if (is.list(word_list)){
     multiple_word_lists=TRUE
     word_list_lengths = sapply(word_list, length)
-    #print(word_list_lengths)
+
     if (all(word_list_lengths == word_list_lengths[1])){
       word_list_length = word_list_lengths[1]
     } else {
       word_list_length = LCM(word_list_lengths)
-      if (list_length > 16^(hash_subsection_size)){
-        if (!suppress_warning) warning(
-          'LCM of provided word lists is greater than range of values.'
+      if (!is.na(forced_limit) & forced_limit < word_list_length){
+        word_list_length=forced_limit
+        if (
+          any(
+            word_list_lengths < forced_limit
+          ) &
+          !suppress_warnings
+        ){
+          warning('The forced_limit value you chose is too small to cover all words in all lists.')
+        }
+      }
+      if (word_list_length > 16^(hash_subsection_size)){
+
+        if (!suppress_warnings) warning(
+          'LCM of provided word lists is greater than range of values. Value is being limited.'
           )
         word_list_length = 16^hash_subsection_size
+      }
+
+      if (hash_output_length < log(word_list_length) & !is.na(hash_output_length)){
+        if (!suppress_warnings){
+          warning('The hash you chose cannot cover all combinations of input.')
+        }
       }
     }
 
@@ -177,10 +244,14 @@ keyToEnglish <- function(
 
   } else {
     multiple_word_lists=FALSE
+    if (hash_output_length < phrase_length * hash_subsection_size){
+      warning('The hash you chose cannot cover all combinations of input.')
+    }
   }
 
+
   if (!is.character(x)){
-    if (!suppress_warning) warning('Converting input to character')
+    if (!suppress_warnings) warning('Converting input to character')
     x = as.character(x)
   }
 
@@ -206,11 +277,12 @@ keyToEnglish <- function(
     original_word_lists = word_list
     word_list = paste0('X',seq_len(word_list_length))
     n_trans_funcs = length(trans_funcs)
+
     original_trans_funcs = trans_funcs
     trans_funcs <- lapply(
       seq_along(word_list_lengths),
       function(i){
-        function(x) original_trans_funcs[[1 + i %% n_trans_funcs ]](
+        function(x) original_trans_funcs[[1 + (i-1) %% n_trans_funcs ]](
           original_word_lists[[i]][
             1 + (as.numeric(gsub('X','', x)) %% word_list_lengths[i])
             ]
@@ -218,7 +290,6 @@ keyToEnglish <- function(
       }
 
     )
-
   }
 
   if (class(hash_function) != 'function')
@@ -229,14 +300,16 @@ keyToEnglish <- function(
     sprintf('.{%s}', hash_subsection_size)
   )
 
-  if (is.na(word_list[1])){
-    word_list = utils::read.csv(corpus_path, header=FALSE)[,1,T]
-  }
   n_words = length(word_list)
 
   seq_idx = seq_len(phrase_length)
 
-  #print(trans_funcs)
+  if (hash_subsection_size <= 7 ){
+    convert <- strtoi
+  } else {
+    # this is slower for longer lists, but it handles values >= 2^32
+    convert <- as.numeric
+  }
 
   if (length(trans_funcs) > 1){
     word_trans_function <- function(x, i){
@@ -246,7 +319,7 @@ keyToEnglish <- function(
     new_keys = unlist(lapply(
       split_hashes,
       function(x) paste(word_trans_function(word_list[
-        1 + strtoi(paste0('0x', x[seq_idx])) %% n_words
+        1 + convert(paste0('0x', x[seq_idx])) %% n_words
       ], seq_idx), collapse=sep)
     ))
   } else {
@@ -255,7 +328,7 @@ keyToEnglish <- function(
       split_hashes,
       function(x) paste(
         word_trans_function(
-          word_list[strtoi(paste0('0x', x[seq_idx])) %% n_words]
+          word_list[convert(paste0('0x', x[seq_idx])) %% n_words]
           ),
         collapse=sep
       )
@@ -265,7 +338,80 @@ keyToEnglish <- function(
   return(new_keys)
 }
 
-# least-common multiple for arbitrary number of arguments
+#' Hash to Sentence
+#'
+#' Hashes data to a sentence that contains 54 bits of entropy
+#'
+#' @export
+#' @param x - Input data, which will be converted to `character` if not already `character`
+#' @param ... - Other parameters to pass to `keyToEnglish()`,
+#'              besides `word_list`, `hash_subsection_size`, and `hash_function`
+#'
+#' @return `character` vector of hashed field resembling phrases
+hash_to_sentence <- function(
+  x,
+  ...
+){
+  keyToEnglish(
+    x,
+    word_list=keyToEnglish::wml_long_sentence,
+    hash_subsection_size=3,
+    hash_function='sha224',
+    ...
+  )
+}
+
+#' Generate Random Sentences
+#'
+#' Randomly generate sentences with a specific structure
+#'
+#' @export
+#' @param n `numeric` number of sentences to generate
+#' @param punctuate `logical` value of whether to add spaces, capitalize first letter, and append period
+#'
+#' @return `character` vector of randomly generated sentences
+generate_random_sentences <- function(
+  n,
+  punctuate=TRUE
+){
+  salt_bytes = openssl::rand_bytes(128)
+  salt_bytes[salt_bytes==0] = as.raw(1)
+  salt = rawToChar(salt_bytes)
+  x = openssl::rand_num(n)
+  x = paste(x, salt)
+
+  if (!punctuate){
+    return(hash_to_sentence(x))
+  } else{
+    capitalizer=list(
+      stringr::str_to_title,
+      identity,
+      identity,
+      identity,
+      identity,
+      identity
+    )
+    return(paste0(
+      hash_to_sentence(
+        x,
+        sep=' ',
+        word_trans=capitalizer
+      ),
+      '.'
+    )
+    )
+  }
+}
+
+#' Least Common Multiple
+#'
+#' Calculates least common multiple of a list of numbers
+#'
+#' @export
+#'
+#' @param ... Any number of `numeric` vectors or nested `list`s containing such
+#'
+#' @return A `numeric` that is the least common multiple of the input values
 LCM <- function(...){
   vals = unlist(list(...))
   n_vals = length(vals)
@@ -277,7 +423,15 @@ LCM <- function(...){
   )
 }
 
-# greatest-common denominator for arbitrary number of arguments
+#' Greated Common Denominator
+#'
+#' Calculates greatest common denominator of a list of numbers
+#'
+#' @export
+#'
+#' @param ... Any number of `numeric` vectors or nested `list`s containing such
+#'
+#' @return A `numeric` that is the greatest common denominator of the input values
 GCD <- function(...){
   vals = unlist(list(...))
   n_vals = length(vals)
@@ -297,11 +451,58 @@ gcd <- function(x, y){
   if (x == y){
     return(x)
   }
-  for (i in ceiling(sqrt(max(x, y))):2){
-    if (x %% i == 0 & y %% i == 0){
-      return(i)
+  v0 = min(x,y)
+  v = v0
+  last_factor = 1
+  while (v > 1){
+    if (x %% v == 0 & y %% v == 0){
+      return(v)
+    }
+    if (v >= v0 %/% last_factor + 1){
+      last_factor = last_factor + 1
+      v = ceiling(v0/last_factor)
+    } else {
+      v = v - 1
     }
   }
   return(1)
+}
+
+#' Uniqueness Probability
+#'
+#' Calculates probability that all `r` elements of a set of size `N` are unique
+#' @export
+#' @param N `numeric` size of set. Becomes unstable for values greater than 10^16.
+#' @param r `numeric` number of elements selected with replacement
+#'
+#' @return `numeric` probability that all `r` elements are unique
+uniqueness_probability <- function(N, r){
+  exp(lgamma(N+1)-lgamma(N - r + 1) - r * log(N))
+}
+
+#' Uniqueness Max Size
+#'
+#' Returns approximate number of elements that you can select out
+#' of a set of size `N` if the probability of there being any duplicates
+#' is less than or equal to `p`
+#' @export
+#'
+#' @param N `numeric` size of set elements are selected from, or a `list` of
+#'          `list`s of `character` vectors (e.g., `wml_animals`)
+#' @param p `numeric` probability that there are any duplicate elements
+#'
+#' @returns `numeric` value indicating size. Value will most likely be non-integer
+#'
+#' @examples
+#' # how many values from 1-1,000 can I randomly select before
+#' # I have a 10% chance of having at least one duplicate?
+#'
+#' uniqueness_max_size(1000,0.1)
+#' # 14.51
+uniqueness_max_size <- function(N, p){
+  if (is.list(N)){
+    N = prod(unlist(lapply(N, length)))
+  }
+  sqrt(2*log(1/(1-p))*N)
 }