+++ NEWS.md
@@ -1,3 +1,23 @@
+# keyToEnglish 0.2.0
+
+## Major changes
+
+* `keyToEnglish()` argument `word_list` can now be a list of word lists, each representing a position in the phrases built. This can be used to construct grammatically-correct phrases
+* Included several curated word lists for more sensical and useful patterns.
+
+
+## Minor changes
+
+* Include warnings when hash function cannot cover all combinations of input
+* Added function `hash_to_sentence` that produces (more or less) grammatically correct phrases with high entropy (54 bits)
+* Added two functions to assist with hash collision probability calculations
+* Added a function to randomly generate sentences as keys.
+* Added least-common-multiple and greatest-common-denominator functions.
+
+## Bugfixes
+
+* Fixed indexing bug in `keyToEnglish()` where an index of 0 would produce an `NA` value. This also changes the values in the tests and the function will not produce the same values as before.
+
 # keyToEnglish 0.1.0
 
 ## Major changes